feat: home 아이콘 추가
