--- components/NavBar/index.tsx
@@ -1,3 +1,0 @@
-export default function NavBar() {
-  return <div>네비게이션 바</div>;
-}
